fix(java): resolve two CodeQL findings in nv-boot starters
- BootWarmupBase: widen totalSuccesses to long. Stream.count() returns long,
  so accumulating it into an int narrowed the value (alert #162).
- OciRegistryClient: make the registry-host regex labels dot-free
  ([^/.]+ separated by literal dots) instead of [^/]+(?:\.[^/]+)*. The old
  form let a label contain '.', so it was ambiguous and backtracked
  exponentially on inputs like '..' repeated (ReDoS, alert #161). The tighter
  form matches the same valid hosts (dotted names, ports, IPs, single labels)
  and rejects the malformed '..'-prefixed inputs.

+++ src/libraries/java/nv-boot-parent/nv-boot-starter-core/src/main/java/com/nvidia/boot/core/warmup/BootWarmupBase.java
@@ -98,7 +98,7 @@ public abstract class BootWarmupBase
     @Override
     public void onApplicationEvent(ApplicationReadyEvent event) {
         var start = Instant.now();
-        int totalSuccesses = 0;
+        long totalSuccesses = 0;
         int total = 0;
         var runnables = this.createWarmupTasks();
         var map = runnables.stream().collect(Collectors.groupingBy(WarmupRunnable::getOrder));
